pyproject.toml
[tool.poetry]
name = "ai-2026"
version = "0.1.0"
description = ""
authors = ["Piotr Marecki <pmarecki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pydantic = "^2.12.5"
loguru = "^0.7.3"
python-dotenv = "^1.2.2"
openai = "^2.29.0"
xai-sdk = "^1.11.0"
anthropic = "^0.96.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
